Fix compilation for 4d0f88e06c6580f3478e5776a6444316ada86ab6

+++ src/common/util.h
@@ -38,7 +38,7 @@ namespace util
 	u32 get_crc32(u8* data, u32 length);
 
 	u8 xbit(u8 a, u8 b);
-	u16 xbit(u16 a, u8 b);
+	u16 xbit(u16 a, u16 b);
 	u32 xbit(u32 a, u32 b);
 
 	extern u32 crc32_table[256];
